pyproject.toml
[project]
name = "genkit-gemini-server"
version = "0.1.1"
description = "High-performance FastAPI server for Gemini API with strict typing and modern tooling"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "gemini-web-wrapper contributors" }]
keywords = ["api", "async", "fastapi", "gemini", "genkit", "type-safe"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Typing :: Typed",
]
dependencies = [
  "aiosqlite>=0.20.0",
  # Cookie management
  "browser-cookie3>=0.20.1",
  "cachetools>=5.3.0",
  # Security fixes for transitive dependencies
  "cryptography>=46.0.0",
  "python-dotenv",
  "fastapi>=0.124.0",
  "gemini-webapi>=1.10.2",
  # Gemini/Genkit dependencies
  "genkit>=0.1.0",
  "genkit-plugin-google-genai>=0.1.0",
  "httpx>=0.28.1",
  "orjson>=3.10.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "sqlalchemy>=2.0.0",
  "uvicorn[standard]>=0.38.0",
  "uvloop>=0.22.0",
  "sqlalchemy",
  "anthropic>=0.30.0",
  "github-copilot-sdk>=0.1.0",
]


[project.urls]
Homepage = "https://gemini-web-wrapper.vercel.app"
Issues = "https://github.com/Ven0m0/gemini-web-wrapper/issues"
Repository = "https://github.com/Ven0m0/gemini-web-wrapper"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
  "test_*.py",
  "example_usage.py",
  ".venv",
  ".git",
  "__pycache__",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "N",  # pep8-naming
  "UP",  # pyupgrade
  "B",  # flake8-bugbear
  "SIM",  # flake8-simplify
  "RUF",  # ruff-specific rules
]
ignore = [
  "B008",  # FastAPI Depends() in defaults is intentional
  "E501",  # Line length soft limit, enforced via formatter
]

[tool.ruff.lint.per-file-ignores]
"dependencies.py" = ["F821"]  # Forward references to AppState
"test_*.py" = ["S101"]  # Assert statements allowed in tests

[tool.mypy]
strict = true
python_version = "3.13"
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
# Pragmatic settings for third-party libraries without complete stubs
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_subclassing_any = false
disallow_any_generics = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--durations=0 -v --strict-markers"
testpaths = [".", "tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.bandit]
exclude_dirs = ["tests", "test_*.py"]
skips = ["B101"]  # Skip assert_used in tests

[dependency-groups]
dev = [
  "mypy>=1.19.1",
  "pip>=25.3",
  "pre-commit>=4.5.0",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "ruff>=0.14.9",
  "setuptools>=80.0.0",
  "types-orjson>=3.6.0",
]
